add es6 javascript
