+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="index.css">
     <title>Document</title>
 </head>
 <body>
@@ -59,6 +59,6 @@
             </div>
         </footer>
 
-<script src="main.js"></script>
+<script type="module" src="index.js"></script>
 </body>
 </html>
