+++ modules/books.js
@@ -1,0 +1,9 @@
+class Books {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+    this.id = Math.floor(Math.random() * 1000000);
+  }
+}
+
+export default Books;
